[Separated] today's reservation into components
