+++ src/pages/reservations/new.tsx
@@ -1,30 +1,32 @@
 import React, { useEffect, useRef, useState } from 'react';
+import { useQuery } from 'react-query';
 import { Button, Navbar, Page } from 'framework7-react';
 import { Objects, PageRouteProps, TimeList } from '@constants';
-import { useQuery } from 'react-query';
-import { getObjects } from '@api';
-import { useSetRecoilState, useRecoilValue } from 'recoil';
+import { useRecoilValue, useRecoilState } from 'recoil';
 import { reservationState, selectedDateState, timeListState } from '@atoms';
+import { getObjects, getResevationsForThatDay } from '@api';
 import useCalendar from '@hooks/useCalendar';
 import { dateFormat } from '@js/utils';
 import Form from '@components/reservations/Form';
-import { getResevationsForThatDay } from '@api';
+import TodayReservations from '@components/reservations/TodayReservation';
 import calImg from '@assets/icons/calendar.png';
 
 const ReservationNewPage = ({ f7router }: PageRouteProps) => {
   const calendarRef = useRef(null);
   const { onPageInit, onPageBeforeRemove } = useCalendar(calendarRef, 'reservation-index-calendar-container');
-  const timeList = useRecoilValue(timeListState);
-  const setTimeList = useSetRecoilState(timeListState);
+
+  const [timeList, setTimeList] = useRecoilState(timeListState);
   const reservationsThisMonth = useRecoilValue(reservationState);
-  const [availableTimeList, setAvailableTimeList] = useState<TimeList[]>([]);
-  const [todayReservations, setTodayReservations] = useState([]);
   const selectedStringDate = useRecoilValue(selectedDateState);
   const selectedDate = dateFormat(new Date(selectedStringDate), 'calendar');
+
+  const [availableTimeList, setAvailableTimeList] = useState<TimeList[]>([]);
+  const [todayReservations, setTodayReservations] = useState([]);
   const [selectedStartTime, setSelectedStartTime] = useState<string>('');
   const [selectedEndTime, setSelectedEndTime] = useState<string>('');
   const [numOfClick, setNumOfClick] = useState<number>(0);
   const [isTimeListReady, setIsTimeListReady] = useState(false);
+
   const { data: timeLists, error } = useQuery<Objects<TimeList>, Error>(
     'timeList',
     getObjects({ model_name: 'time_list' }),
@@ -71,18 +73,6 @@ const ReservationNewPage = ({ f7router }: PageRouteProps) => {
     return selectedEndTime >= endTime && selectedStartTime <= startTime;
   }
 
-  function giveColor(i) {
-    const colors = [
-      'bg-red-400',
-      'bg-yellow-300',
-      'bg-green-400',
-      'bg-teal-400',
-      'bg-sky-400',
-      'bg-indigo-400',
-      'bg-purple-400',
-    ];
-    return colors[i % 7];
-  }
   async function checkAvailableReservationTimes() {
     const dateobj = { date: dateFormat(calendarRef.current.value[0], 'day') };
     const reservations = await getResevationsForThatDay(dateobj);
@@ -135,25 +125,7 @@ const ReservationNewPage = ({ f7router }: PageRouteProps) => {
       <Navbar noHairline innerClassName="bg-white" title="연습실 예약" />
       <div id="reservation-index-calendar-container"></div>
       <div className="w-full bg-gray-100 my-2" style={{ height: '2px' }}></div>
-      <ul className="my-5 min-h-20 flex justify-center items-center">
-        {todayReservations.length > 0 ? (
-          todayReservations.map((reservation, idx) => (
-            <li className="w-full px-3 py-2 relative">
-              <div className={`absolute left-0 top-0 w-2 h-full ${giveColor(idx)}`}></div>
-              <div className="ml-2 flex justify-between items-center">
-                <span>{reservation.note}</span>
-                <span className="text-gray-500">{`${dateFormat(reservation.start_at, 'onlyTime')} ~ 
-              ${dateFormat(reservation.end_at, 'onlyTime')}`}</span>
-              </div>
-            </li>
-          ))
-        ) : (
-          <li className="flex flex-col items-center">
-            <span>공연날짜가 다가오고 있어요!</span>
-            <span>연습해야죠 여러분!!!</span>
-          </li>
-        )}
-      </ul>
+      <TodayReservations todayReservations={todayReservations} />
       <div className="w-full bg-gray-100 my-2" style={{ height: '2px' }}></div>
       <div className="ml-4 mt-5 flex items-center">
         <img src={calImg} width="20" />
